implement solution using custom widget for editor
